{{cookiecutter.app_name}}/requirements-dev.txt
flake8
flask-resty-swagger
pytest
pytest-cov
